+++ src/service/connect/protocol/Connect.Send.go
@@ -1,8 +1,8 @@
 package protocol
 
 type SendRequest struct {
-	Clients []uint `json:c`
-	Data    []byte `json:d`
+	Clients []uint64 `json:c`
+	Data    []byte   `json:d`
 }
 
 type SendResponse int
